+++ utils.js
@@ -1,0 +1,11 @@
+export const getRandom = num => {
+	return Math.ceil(Math.random() * num)
+}
+
+export const createElement = (tag, className) => {
+	const $tag = document.createElement(tag)
+
+	if (className) $tag.classList.add(className)
+
+	return $tag
+}
